+++ client/joinus_client/src/hooks/useAppData.js
@@ -1,8 +1,10 @@
 import { useEffect, useState } from "react";
 import { fetchAPI, addData } from "../api";
+import { reactLocalStorage } from "reactjs-localstorage";
 
 export default function useAppData() {
   const [eventsData, setEventsData] = useState([]); //api for all events
+  const [user, setUser] = useState({ id: null, name: null, picture: null });
   const [categoriesData, setCategoriesData] = useState([]); //api for all categories
   const [usersData, setUsersData] = useState([]); // api for all users
   const [joinedEvents, setJoinedEvents] = useState([]); //api for all joined events
@@ -16,7 +18,6 @@ export default function useAppData() {
       fetchAPI("users/user_events"),
     ])
       .then((all) => {
-        console.log(`ALL: ${all[0]}`);
         setEventsData((prev) => [...all[0].data]);
         setCategoriesData((prev) => [...all[1].data]);
         setUsersData((prev) => [...all[2].data]);
@@ -29,5 +30,31 @@ export default function useAppData() {
       });
   }, [reload]);
 
-  return { eventsData, categoriesData, usersData, joinedEvents, setReload, reload };
+  const login = (userID) => {
+    fetchAPI(`user/${userID}`)
+      .then((data) => {
+        setUser((prev) => JSON.stringify(data.data[0]));
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  const logout = () => {
+    setUser({});
+    reactLocalStorage.remove("currentUser");
+  };
+
+  return {
+    eventsData,
+    categoriesData,
+    usersData,
+    joinedEvents,
+    setReload,
+    reload,
+    login,
+    logout,
+    user,
+    setUser,
+  };
 }
